xserver fontconfig mesa git update Nov 24 2017
xserver          commit 559954aaa8d811a22cf918cc16a7d618e12201a0
fontconfig       commit 64895e719dd8d18c52a31d66cd189915bc8c00b8
mesa             commit 35548cae93513875dcad9cc88589add5fac7d836

+++ mesalib/src/gallium/auxiliary/util/u_async_debug.c
@@ -46,7 +46,7 @@ u_async_debug_message(void *data, unsigned *id, enum pipe_debug_type type,
 
    simple_mtx_lock(&adbg->lock);
    if (adbg->count >= adbg->max) {
-      unsigned new_max = MAX2(16, adbg->max * 2);
+      size_t new_max = MAX2(16, adbg->max * 2);
 
       if (new_max < adbg->max ||
           new_max > SIZE_MAX / sizeof(*adbg->messages)) {
